refactor(workbench-shell): ♻️ migrate dashboard sidebar state to threadStore
Remove prop-drilled workspace and thread state from DashboardSidebar by subscribing to threadStore directly. This simplifies the component interface and centralizes state management.

Eliminate unused helper functions `buildInitialWorkspaceThreadDisplayCounts`, `buildInitialWorkspaceThreadHasMore`, `mapRunStateToWorkbenchThreadStatus`, and `mapRunFinishedStatusToThreadStatus` since the status conversion is now handled via `threadRunStatusToDisplayStatus`.
